09042022 with recipes_limit

+++ foodgram/api/views.py
@@ -137,10 +137,11 @@ class SubscriptionViewSet(CreateDestroyMixin):
     #     self.perform_destroy(instance)
     #     return Response(status=status.HTTP_204_NO_CONTENT)
     def perform_destroy(self, instance):
-        author = get_object_or_404(User, id=self.kwargs.get('author_id'))
+        author = get_object_or_404(User, id=int(self.kwargs.get('author_id')))
         user = self.request.user
         instance = get_object_or_404(Follow, author=author, user=user)
-        instance.delete()
+        if instance.user == user:
+            instance.delete()
 
 
 class ShoppingListViewSet(CustomShoppingFavoriteMixin):
